stage/kaiti/完善PPT、撰写发言稿及答辩提问清单/breast_cancer_diagnosis/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.7.0

# 机器学习工具
scikit-learn>=1.2.0
scipy>=1.10.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 进度条
tqdm>=4.65.0

# 其他工具
pathlib>=1.0.1
